Re-engineering the Options structure. All Options are moved to the engine_dispatcher interface and can be stored in database. Some minor code-style fixes.

+++ webEngine/sources/weScan.cpp
@@ -38,7 +38,7 @@ ScanInfo::ScanInfo()
 ScanInfo::~ScanInfo()
 {
     // nothing special at this moment
-    for (int i = 0; i < scan_data.size(); i++) {
+    for (size_t i = 0; i < scan_data.size(); i++) {
         delete scan_data[i];
     }
     scan_data.clear();
